+++ lib/src/cluster.dart
@@ -12,12 +12,16 @@ class Cluster<T> {
             markers.fold<double>(0.0, (p, c) => p + c.location.longitude) /
                 markers.length);
 
+  /// Retrieve cluster items
   Iterable<T> get items => markers.map((m) => m.item);
 
+  /// Get number of clustered items
   int get count => markers.length;
 
+  /// True if cluster is not a single item cluster
   bool get isMultiple => markers.length > 1;
 
+  /// Basic cluster marker id
   String getId() {
     return location.latitude.toString() +
         "_" +