+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.2.0
+
+- Add `stopClusteringZoom` variable
+- Update to `google_maps_flutter` version 1.0.2
+- Improve `extraPercent` calculation (thanks to @buntagonalprism)
+
 ## 0.1.0
 
 - Fix `getMarkers` signature